# Conflicts:
#	src/it/unipv/insfw23/TicketWave/modelController/controller/payment/PaymentDataPController.java

+++ src/it/unipv/insfw23/TicketWave/modelDomain/user/Customer.java
@@ -95,7 +95,7 @@ public class Customer extends User {
         return ticketsList;
     }
 
-    public double getPoints() {
+    public int getPoints() {
 
         return points;
     }
